third_party/blink/web_tests/fast/dom/Window/window-methods-construct.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
  test(t => {
    assert_throws_js(TypeError, function() { new window.blur(); });
  }, "Construct call to window methods throws.");
</script>
